% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bern_plot.R
\name{bern_plot}
\alias{bern_plot}
\title{Bernoulli Density Plot}
\usage{
bern_plot(p)
}
\arguments{
\item{p}{double}
}
\value{
plot
}
\description{
Plot a simple density plot to illustrate the Bernoulli distribution for a
specified probability of success of a Bernoulli event.
}
\examples{
bern_plot(0.7)
}
