% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_plot.R
\name{binom_plot}
\alias{binom_plot}
\title{Binomial Probability Density Plot}
\usage{
binom_plot(n, p)
}
\arguments{
\item{n}{integer}

\item{p}{double}
}
\value{
plot
}
\description{
Plot a simple density plot to illustrate the binomial distribution for a
specified number of trials and a constant probability of success.
}
\examples{
binom_plot(8, 0.5)
}
