% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact_twosamp.R
\name{exact_twosamp_dist}
\alias{exact_twosamp_dist}
\title{Two-Sample Exact Test Distribution}
\usage{
exact_twosamp_dist(y, n, duplicate = TRUE)
}
\arguments{
\item{y}{vector of scores}

\item{n}{sample size of first sample}

\item{duplicate}{Boolean to indicate inclusion of duplicates of statistic}
}
\value{
outputs exact distribution
}
\description{
Two-Sample Exact Test Distribution
}
\examples{
scores <- c(4, 5, 4, 7, 9, 10, 12, 6)
exact_twosamp_dist(scores, 3, duplicate = FALSE)
}
