% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_dist.R
\name{mult_dist}
\alias{mult_dist}
\title{Multinomial Distribution}
\usage{
mult_dist(x, pd, chi2.sort = FALSE)
}
\arguments{
\item{x}{integer vector}

\item{pd}{numeric vector}

\item{chi2.sort}{boolean}
}
\value{
data frame of probability distribution
}
\description{
Multinomial Distribution
}
\examples{
observed <- c(5, 3, 0, 0)
probabilities <- c(0.25, 0.25, 0.25, 0.25)
mult_dist(observed, probabilities)
}
