% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_scores.R
\name{norm_scores}
\alias{norm_scores}
\title{Normal Scores Generation}
\usage{
norm_scores(x, two.sample = FALSE)
}
\arguments{
\item{x}{double with scores}

\item{two.sample}{Boolean indicating one or two samples}
}
\value{
double with normal scores
}
\description{
Calculate normal scores for a single sample.
}
\examples{
scores <- c(5, 9, -3, 8, -7)
norm_scores(scores)
}
