% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nplearn.R
\docType{package}
\name{nplearn-package}
\alias{nplearn}
\alias{nplearn-package}
\title{nplearn: Nonparametric Statistics Learning}
\description{
This is to facilitate the teaching and learning of nonparametric
    statistics. The functions in the package are primarily for explaining
    concepts, rather than for real-world application.
}
\author{
\strong{Maintainer}: Michael Seaman \email{fourthz@gmail.com} (0000-0002-7406-5398)

}
\keyword{internal}
