% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_plot.R
\name{power_plot}
\alias{power_plot}
\title{Power Plot}
\usage{
power_plot(h_0, h_1, alpha = 0.05, line = 0)
}
\arguments{
\item{h_0}{double}

\item{h_1}{double}

\item{alpha}{double}

\item{line}{double}
}
\value{
plot
}
\description{
Create a plot of power as a function of sample size for an upper-tail
binomial exact test using simple null and alternative hypotheses,
a maximum Type I error rate, and an optional power target.
}
\examples{
power_plot(h_0 = 0.5, h_1 = 0.6, alpha = .05, line = 0.80)
}
