% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prodbinom_dist.R
\name{prodbinom_dist}
\alias{prodbinom_dist}
\title{Product Binomial Distribution}
\usage{
prodbinom_dist(n1, n2, p1, p2, two.sided = TRUE)
}
\arguments{
\item{n1}{integer}

\item{n2}{integer}

\item{p1}{double}

\item{p2}{double}
}
\value{
data frame of probability distribution
}
\description{
Product Binomial Distribution
}
\examples{
prodbinom_dist(5, 3, .5, .4, two.sided = TRUE)
}
