% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_dist.R
\name{rand_dist}
\alias{rand_dist}
\title{Randomization Test Distribution}
\usage{
rand_dist(x)
}
\arguments{
\item{x}{double}
}
\value{
output
}
\description{
Display the distribution of the test statistic for a single-sample
randomization test.
}
\examples{
scores <- (5, 3, -7)
rand_dist(scores)
}
