% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_scores.R
\name{norm_scores}
\alias{norm_scores}
\title{Normal Scores Generation}
\usage{
norm_scores(x)
}
\arguments{
\item{x}{double}
}
\value{
double
}
\description{
Calculate normal scores for a single sample.
}
\examples{
scores <- c(5, 9, -3, 8, -7)
norm_scores(scores)
}
